scripts/requirements.txt
aniso8601==9.0.1
Brotli==1.0.9
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.1.2
colorama==0.4.4
dash==2.3.1
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
Flask==2.1.1
Flask-Caching==1.10.1
Flask-Compress==1.11
Flask-RESTful==0.3.9
gunicorn==20.1.0
idna==3.3
importlib-metadata==4.11.3
itsdangerous==2.1.2
Jinja2==3.1.1
MarkupSafe==2.1.1
numpy==1.22.3
packaging==21.3
panda==0.3.1
pandas==1.4.2
patsy==0.5.2
pip==21.2.4
plotly==5.7.0
psycopg2-binary==2.9.3
pyparsing==3.0.8
python-dateutil==2.8.2
pytz==2022.1
requests==2.27.1
scipy==1.8.0
setuptools==61.2.0
six==1.16.0
statsmodels==0.13.2
tenacity==8.0.1
urllib3==1.26.9
Werkzeug==2.1.1
wheel==0.37.1
wincertstore==0.2
zipp==3.8.0
